Clean code

+++ Peak_UB_LB/I_SAT.py
@@ -72,8 +72,6 @@ def read_input(filename):
                     else:
                         break
                 cnt = cnt + 1
-    for i in range(n):
-        delv(i, temp)
         
     with open('task_power/' + filename + '.txt', 'r') as k:
         W = [int(line.strip()) for line in k if line.strip()]
@@ -99,33 +97,6 @@ def reset(idx):
     adj = []
     ran = []
     forward = [0 for i in range(200)]
-
-
-'''
-    E** = E + (u, v) if u in E and v in E and (u, v) not in E
-    - Nếu u là đỉnh gốc (root_i), thì append (u, k) vào adj nếu chưa có
-    - Logic cập nhật temp[i] vẫn giữ nguyên để các tầng trên có dữ liệu quy hoạch động
-'''
-def delv(i, temp, root_i=None):
-    global adj, neighbors, reversed_neighbors, ran
-    if root_i is None:
-        root_i = i
-    if len(temp[i]) == 0:
-        return []
-    if ran[i] == 1:
-        return temp[i]
-        
-    for j in temp[i]:
-        con = delv(j, temp, root_i)
-        if con:
-            for k in con:
-                if i == root_i: 
-                    if [i, k] not in adj:
-                        adj.append([i, k])
-                        neighbors[i][k] = 1
-                        reversed_neighbors[k][i] = 1
-    ran[i] = 1
-    return temp[i]
 
 def generate_variables(n,m,c):
     global var_counter
@@ -437,7 +408,7 @@ def get_value(solution, c):
 
         return value, [table[i][:value] for i in range(m)]
 
-def optimal(X, S, A, n, m, makespan, sol, start_time, peak):
+def optimal(X, S, A, n, m, makespan, sol, start_time, peak, lower_bound):
     global filename
     
     ip1, ip2 = preprocess(n, m, makespan, time_list, adj)
@@ -460,6 +431,9 @@ def optimal(X, S, A, n, m, makespan, sol, start_time, peak):
     print("New makespan:", bestValue, end="\r")
 
     while (True):
+        if bestValue - 1 < lower_bound:
+            print("Reached lower bound. Optimal solution found.")
+            return bestValue, ansmap, var_counter, clauses, "Optimal", sol
         solver.delete()
         solver = Cadical195()
         clauses = generate_clauses(n, m, bestValue - 1, time_list, adj, ip1, ip2, X, S, A, peak)
@@ -498,7 +472,7 @@ def calculate_peak():
     peak = upper_lower_power_cap(W, m)
     lower_bound = analytical_cycle_lower_bound(time_list, W, m, peak)
     makespan = initial_probe_horizon(time_list, lower_bound, m)
-    return peak, makespan
+    return peak, makespan, lower_bound
 
 if __name__ == "__main__":
     filename = sys.argv[1]
@@ -507,9 +481,11 @@ if __name__ == "__main__":
     read_input(filename)
     sol = 0
     startime = time.time()
-    peak, makespan = calculate_peak()
+    peak, makespan, lower_bound = calculate_peak()
     X, A, S = generate_variables(n,m,makespan)
-    bestValue, ansmap, var, clauses, status, sol = optimal(X,S,A,n,m,makespan,sol,startime, peak)
+    bestValue, ansmap, var, clauses, status, sol = optimal(
+        X, S, A, n, m, makespan, sol, startime, peak, lower_bound
+    )
     for line in ansmap:
         print(line)
     endtime = time.time()
